requirements-dev.txt
asgiref==3.6.0
autopep8==2.0.2
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.1.7
django-restframework==0.0.1
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
static3==0.7.0
tomli==2.0.1
